- File input channel waits rather than shuffling to disk if memory storage is full. - BinarySize handles "G" format string as required by IFormattable.
git-svn-id: file:///mnt/c/Users/svgroot/Downloads/svn/repos/Jumbo/branches/oldfilechannel@853 2f0f0f96-4b5b-11dd-99f4-89bc9257463c

+++ Tkl.Jumbo.Test/UnmanagedBufferTests.cs
@@ -19,7 +19,7 @@ namespace Tkl.Jumbo.Test
                 byte[] expected = Utilities.GenerateData(512);
                 byte[] actual = new byte[512];
 
-                int index = UnmanagedBuffer.CopyCircular(expected, 0, target, 0, expected.Length);
+                long index = UnmanagedBuffer.CopyCircular(expected, 0, target, 0, expected.Length);
 
                 Assert.AreEqual(512, index);
                 UnmanagedBuffer.CopyCircular(target, 0, actual, 0, 512);
